Conflicts:
	algo/tools/align.cpp

+++ server/web2py/applications/vidjil/models/task.py
@@ -1,13 +1,32 @@
 # coding: utf8
 import os
-import sys
 import defs
 import re
+import os.path
+import time
+import sys
+import datetime
+import random
+import xmlrpclib
 
-TASK_TIMEOUT = 60 * 60
 
-def schedule_run(id_sequence, id_config):
-    import time, datetime, sys, os.path
+def assert_scheduler_task_does_not_exist(args):
+    ##check scheduled run
+    row = db( ( db.scheduler_task.args == args)
+         & ( db.scheduler_task.status != "FAILED"  )
+         & ( db.scheduler_task.status != "EXPIRED"  )
+         & ( db.scheduler_task.status != "TIMEOUT"  )
+         & ( db.scheduler_task.status != "COMPLETED"  )
+         ).select()
+
+    if len(row) > 0 :
+        res = {"message": "task already registered"}
+        return res
+
+    return None
+
+
+def schedule_run(id_sequence, id_config, grep_reads=None):
     from subprocess import Popen, PIPE, STDOUT, os
 
     id_patient = db.sequence_file[id_sequence].patient_id
@@ -22,48 +41,44 @@ def schedule_run(id_sequence, id_config):
                                      run_date = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S'),
                                      config_id = id_config )
         
-    ## check fused_file
-    row2 = db( ( db.fused_file.config_id == id_config ) & 
-              ( db.fused_file.patient_id == id_patient )  
-            ).select()
+    if grep_reads:
+        args = [id_sequence, id_config, data_id, None, grep_reads]
+    else:
+        ## check fused_file
+        row2 = db( ( db.fused_file.config_id == id_config ) &
+                   ( db.fused_file.patient_id == id_patient )
+               ).select()
 
-    if len(row2) > 0 : ## update
-        fuse_id = row2[0].id
-    else:             ## create
-        fuse_id = db.fused_file.insert(patient_id = id_patient,
-                                        config_id = id_config)
-        
-    ##check scheduled run
-    row3 = db( ( db.scheduler_task.args == '["' + str(id_sequence) + '", "' + str(id_config) + '", ' + str(data_id) + ', ' + str(fuse_id) + ']' ) 
-         & ( db.scheduler_task.status != "FAILED"  )
-         & ( db.scheduler_task.status != "EXPIRED"  )
-         & ( db.scheduler_task.status != "TIMEOUT"  )
-         & ( db.scheduler_task.status != "COMPLETED"  )
-         ).select()
+        if len(row2) > 0 : ## update
+            fuse_id = row2[0].id
+        else:             ## create
+            fuse_id = db.fused_file.insert(patient_id = id_patient,
+                                           config_id = id_config)
 
-    if len(row3) > 0 :
-        res = {"message": "run already registered"}
-        log.error(res)
-        return res
+        args = [id_sequence, id_config, data_id, fuse_id, None]
 
+    err = assert_scheduler_task_does_not_exist(str(args))
+    if err:
+        log.error(err)
+        return err
 
     program = db.config[id_config].program
     ##add task to scheduler
-    task = scheduler.queue_task(program, [id_sequence, id_config, data_id, fuse_id]
-                                , repeats = 1, timeout = TASK_TIMEOUT)
+    task = scheduler.queue_task(program, args,
+                                repeats = 1, timeout = defs.TASK_TIMEOUT)
     db.results_file[data_id] = dict(scheduler_task_id = task.id)
 
     filename= db.sequence_file[id_sequence].filename
 
     res = {"redirect": "reload",
-           "message": "[%s] (%s) c%s: process requested - %s" % (data_id, id_patient, id_config, filename)}
+           "message": "[%s] (%s) c%s: process requested - %s %s" % (data_id, id_patient, id_config, grep_reads, filename)}
 
     log.info(res)
     return res
 
 
-def run_vidjil(id_file, id_config, id_data, id_fuse, clean_before=False, clean_after=False):
-    import time, datetime, sys, os.path
+def run_vidjil(id_file, id_config, id_data, id_fuse, grep_reads,
+               clean_before=False, clean_after=False):
     from subprocess import Popen, PIPE, STDOUT, os
     
     ## les chemins d'acces a vidjil / aux fichiers de sequences
@@ -83,7 +98,11 @@ def run_vidjil(id_file, id_config, id_data, id_fuse, clean_before=False, clean_a
 
     ## config de vidjil
     vidjil_cmd = db.config[id_config].command
-    vidjil_cmd = vidjil_cmd.replace( 'germline/' ,germline_folder)
+    vidjil_cmd = vidjil_cmd.replace( ' germline' ,germline_folder)
+
+    if grep_reads:
+        # TODO: security, assert grep_reads XXXX
+        vidjil_cmd += ' -FaW "%s" ' % grep_reads
     
     os.makedirs(out_folder)
     out_log = out_folder+'/'+output_filename+'.vidjil.log'
@@ -107,13 +126,19 @@ def run_vidjil(id_file, id_config, id_data, id_fuse, clean_before=False, clean_a
     sys.stdout.flush()
     db.commit()
 
-    ## récupération du fichier data.json généré
-    results_filepath = os.path.abspath(out_folder+'/'+output_filename+".vidjil")
+    ## Get result file
+    if grep_reads:
+        out_results = out_folder + '/seq/clone.fa-1'
+    else:
+        out_results = out_folder + '/' + output_filename + '.vidjil'
+
+    print "===>", out_results
+    results_filepath = os.path.abspath(out_results)
 
     try:
         stream = open(results_filepath, 'rb')
     except IOError:
-        print "!!! Vidjil failed, no .vidjil file"
+        print "!!! Vidjil failed, no result file"
         res = {"message": "[%s] c%s: Vidjil FAILED - %s" % (id_data, id_config, out_folder)}
         log.error(res)
         raise IOError
@@ -168,14 +193,15 @@ def run_vidjil(id_file, id_config, id_data, id_fuse, clean_before=False, clean_a
     res = {"message": "[%s] c%s: Vidjil finished - %s - %s" % (id_data, id_config, info, out_folder)}
     log.info(res)
 
-    run_fuse(id_file, id_config, id_data, id_fuse, clean_before = False)
+
+    if not grep_reads:
+        run_fuse(id_file, id_config, id_data, id_fuse, clean_before = False)
 
     return "SUCCESS"
 
 
 
 def run_copy(id_file, id_config, id_data, id_fuse, clean_before=False, clean_after=False):
-    import time, datetime, sys, os.path
     from subprocess import Popen, PIPE, STDOUT, os
     
     ## les chemins d'acces a vidjil / aux fichiers de sequences
@@ -237,7 +263,6 @@ def run_copy(id_file, id_config, id_data, id_fuse, clean_before=False, clean_aft
 
 
 def run_fuse(id_file, id_config, id_data, id_fuse, clean_before=True, clean_after=False):
-    import time, datetime, sys, os.path
     from subprocess import Popen, PIPE, STDOUT, os
     
     out_folder = defs.DIR_OUT_VIDJIL_ID % id_data
@@ -315,11 +340,12 @@ def run_fuse(id_file, id_config, id_data, id_fuse, clean_before=True, clean_afte
     return "SUCCESS"
 
 def custom_fuse(file_list):
-    import time, datetime, sys, os.path, random, xmlrpclib
     from subprocess import Popen, PIPE, STDOUT, os
-    
+
+    if defs.PORT_FUSE_SERVER is None:
+        raise IOError('This server cannot fuse custom data')
     random_id = random.randint(99999999,99999999999)
-    out_folder = defs.DIR_OUT_VIDJIL_ID % random_id
+    out_folder = os.path.abspath(defs.DIR_OUT_VIDJIL_ID % random_id)
     output_filename = defs.BASENAME_OUT_VIDJIL_ID % random_id
     
     cmd = "rm -rf "+out_folder 
@@ -335,7 +361,7 @@ def custom_fuse(file_list):
     files = ""
     for id in file_list :
         if db.results_file[id].data_file is not None :
-            files += defs.DIR_RESULTS + db.results_file[id].data_file + " "
+            files += os.path.abspath(defs.DIR_RESULTS + db.results_file[id].data_file) + " "
     
     cmd = "python "+ os.path.abspath(defs.DIR_FUSE) +"/fuse.py -o "+output_file+" -t 100 "+files
     proc_srvr = xmlrpclib.ServerProxy("http://localhost:%d" % defs.PORT_FUSE_SERVER)
@@ -344,10 +370,10 @@ def custom_fuse(file_list):
     try:
         f = open(fuse_filepath, 'rb')
         data = gluon.contrib.simplejson.loads(f.read())
-    except IOError:
+    except IOError, e:
         res = {"message": "'custom fuse' -> IOError"}
         log.error(res)
-        raise IOError
+        raise e
 
     clean_cmd = "rm -rf " + out_folder 
     p = Popen(clean_cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
